requirements.txt
youtube-transcript-api==0.6.2
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
pytube==15.0.0
pytube3==9.6.4
Pillow==9.0.0
numpy==1.24.4
Flask==3.0.0
